pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "twitcast"
version = "0.1.0"
description = "TWiT network podcast tools: dashboard, promo, summarization"
requires-python = ">=3.12"
dependencies = [
    "requests",
    "Pillow",
    "click",
    "anthropic",
]

[project.scripts]
twitcast = "twitcast.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
